backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
youtube-transcript-api==0.6.1
openai==0.28.1
python-dotenv==1.0.0
